Fixed test naming for onell * removed not used imports * deleted unused local variables
